Removed NewSoftSerial debug and added a "ping" after node detection.
The Zigbee/ZNet protocol used by the XBee is quite advanced and I noticed the first packet sent took longer than any subsequent package. Seconds longer. Haven't dig into it but my guess is that it's doing a check to find best route. So to give it a chance we "ping" the fireCrow by sending a payload of 254, which the fireCrow firmware just ignores, but the XBees gets to do their thing.